#44: Uploaden van een scan
Refactor, logica uit het Screen naar een hook verplaatst
